Inlined healthHelpers.mjs into HealthService.mjs
